backend/requirements.txt
asyncpg
bcrypt 
fastapi
fastexcel
gunicorn
libsql-client
sqlalchemy
polars
pydantic
python-dotenv
uvicorn
jwt
passlib